Move fetchGitCredentials into git_creds.go
