+++ pkg/controller/gittrack/gittrack_controller.go
@@ -208,56 +208,6 @@ func (r *ReconcileGitTrack) checkoutRepo(url string, ref string, gitCreds *gitCr
 	return repo, nil
 }
 
-// fetchGitCredentials creates git credentials data from a given deployKey secret reference
-func (r *ReconcileGitTrack) fetchGitCredentials(gt farosv1alpha1.GitTrackInterface) (*gitCredentials, error) {
-	deployKey := gt.GetSpec().DeployKey
-
-	// Check if the deployKey is empty, do nothing if it is
-	emptyKey := farosv1alpha1.GitTrackDeployKey{}
-	if deployKey == emptyKey {
-		return nil, nil
-	}
-	// Check the deployKey fields are both non-empty
-	if deployKey.SecretName == "" || deployKey.Key == "" {
-		return nil, fmt.Errorf("if using a deploy key, both SecretName and Key must be set")
-	}
-
-	// Set the SecretNamespace to match the GitTrack Namespace
-	// Users must set this for ClusterGitTracks
-	switch gt.(type) {
-	case *farosv1alpha1.GitTrack:
-		if deployKey.SecretNamespace != "" && deployKey.SecretNamespace != gt.GetNamespace() {
-			return nil, fmt.Errorf("DeployKey namespace must match GitTrack namespace or be empty")
-		}
-		deployKey.SecretNamespace = gt.GetNamespace()
-	case *farosv1alpha1.ClusterGitTrack:
-		if deployKey.SecretNamespace == "" {
-			return nil, fmt.Errorf("No Secret Namespace set for DeployKey")
-		}
-	default:
-		panic(fmt.Errorf("This code should not be reachable"))
-	}
-
-	// Fetch the secret from the API
-	secret := &apiv1.Secret{}
-	err := r.Get(context.TODO(), types.NamespacedName{
-		Namespace: deployKey.SecretNamespace,
-		Name:      deployKey.SecretName,
-	}, secret)
-	if err != nil {
-		return nil, fmt.Errorf("failed to look up secret %s: %v", deployKey.SecretName, err)
-	}
-
-	// Extract the data from the secret
-	var ok bool
-	var secretData []byte
-	if secretData, ok = secret.Data[deployKey.Key]; !ok {
-		return nil, fmt.Errorf("invalid deploy key reference. Secret %s does not have key %s", deployKey.SecretName, deployKey.Key)
-	}
-
-	return &gitCredentials{secret: secretData, credentialType: deployKey.Type}, nil
-}
-
 // getFiles checks out the Spec.Repository at Spec.Reference and returns a map of filename to
 // gitstore.File pointers
 func (r *ReconcileGitTrack) getFiles(gt farosv1alpha1.GitTrackInterface) (map[string]*gitstore.File, error) {
